Add UiText class
This construct defers the resolution of the string value to the
composition time or whenever it is requested as opposed to when it is
set up.

Using this allows business logic further down the stack to pass and
handle strings back up without the need to know Android Context or
Resources.